Revert "RGBEyes bugfixes"
This reverts commit af7bec7382d733392ae1553312169a15421c7f80.

+++ include/controllers/RGBEyesController.h
@@ -84,7 +84,7 @@ namespace bjos {
 		void matrix_render(void);
 		void matrix_raise(void);
 		void matrix_bottom(void);
-		uint32_t createRGB(int r, int g, int b);
+		unsigned long createRGB(int r, int g, int b);
       
         /* Initialize the main instance */ 
         void init(BJOS *bjos);
